Rename NS XmlDocumentCleaner to XmlDocumentCleaners
