+++ src/XmlDocumentCleaners/SetKnownSchemaLocations.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PhpCfdi\CfdiCleaner\XmlDocumentCleaner;
+namespace PhpCfdi\CfdiCleaner\XmlDocumentCleaners;
 
 use DOMAttr;
 use DOMDocument;
