+++ tests/Features/XmlDocumentCleaners/MoveNamespaceDeclarationToRootTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace PhpCfdi\CfdiCleaner\Tests\Features\XmlDocumentCleaner;
+namespace PhpCfdi\CfdiCleaner\Tests\Features\XmlDocumentCleaners;
 
 use PhpCfdi\CfdiCleaner\Tests\TestCase;
-use PhpCfdi\CfdiCleaner\XmlDocumentCleaner\MoveNamespaceDeclarationToRoot;
+use PhpCfdi\CfdiCleaner\XmlDocumentCleaners\MoveNamespaceDeclarationToRoot;
 
 final class MoveNamespaceDeclarationToRootTest extends TestCase
 {
