+++ tests/Features/XmlDocumentCleaners/MoveSchemaLocationsToRootTest.php
@@ -4,10 +4,10 @@
 
 declare(strict_types=1);
 
-namespace PhpCfdi\CfdiCleaner\Tests\Features\XmlDocumentCleaner;
+namespace PhpCfdi\CfdiCleaner\Tests\Features\XmlDocumentCleaners;
 
 use PhpCfdi\CfdiCleaner\Tests\TestCase;
-use PhpCfdi\CfdiCleaner\XmlDocumentCleaner\MoveSchemaLocationsToRoot;
+use PhpCfdi\CfdiCleaner\XmlDocumentCleaners\MoveSchemaLocationsToRoot;
 
 final class MoveSchemaLocationsToRootTest extends TestCase
 {
