+++ tests/Features/XmlDocumentCleaners/RemoveUnusedNamespacesTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace PhpCfdi\CfdiCleaner\Tests\Features\XmlDocumentCleaner;
+namespace PhpCfdi\CfdiCleaner\Tests\Features\XmlDocumentCleaners;
 
 use PhpCfdi\CfdiCleaner\Tests\TestCase;
-use PhpCfdi\CfdiCleaner\XmlDocumentCleaner\RemoveUnusedNamespaces;
+use PhpCfdi\CfdiCleaner\XmlDocumentCleaners\RemoveUnusedNamespaces;
 
 final class RemoveUnusedNamespacesTest extends TestCase
 {
